Removed <a> tag from download button
The <a> tag was overriding the button functionality. If the user clicks the background of the button and not the text the function worked as intended, but this was counter intuitive so I have removed the hyperlink for now.

If we want to link to a page which displays the pdf there might be a way to do this Javascript instead of using a hyperlink.