Add unit tracking from introduction
This works for the Master of Translation, but my concern is that it will not be robust if the description of unit points to be taken is not consistent in the introduction descriptions.

Perhaps this can be overcome with more sophisticated regex searches, however ultimately I think it would be most beneficial to implement some standards within the data source.